Subscriber.tg_chat_id: IntegerField -> BigIntegerField (#584)

+++ apps/bot_init/models.py
@@ -50,7 +50,7 @@ class AdminMessage(models.Model):
 class Subscriber(models.Model):
     """Модель подписчика бота."""
 
-    tg_chat_id = models.IntegerField(verbose_name='Идентификатор подписчика', unique=True)
+    tg_chat_id = models.BigIntegerField(verbose_name='Идентификатор подписчика', unique=True)
     is_active = models.BooleanField(default=True, verbose_name='Подписан ли пользователь на бота')
     step = models.CharField(max_length=100, verbose_name='Шаг пользователя', blank=True, null=True)
     comment = models.TextField(null=True, blank=True, verbose_name='Комментарий к подписчику')
@@ -92,9 +92,9 @@ class Message(models.Model):
     """Модель для хранения сообщений."""
 
     date = models.DateTimeField(null=True, verbose_name='Дата отправки')
-    from_user_id = models.IntegerField(verbose_name='Идентификатор отправителя')
-    message_id = models.IntegerField(verbose_name='Идентификатор сообщения')
-    chat_id = models.IntegerField(verbose_name='Идентификатор чата, в котором идет общение')
+    from_user_id = models.BigIntegerField(verbose_name='Идентификатор отправителя')
+    message_id = models.BigIntegerField(verbose_name='Идентификатор сообщения')
+    chat_id = models.BigIntegerField(verbose_name='Идентификатор чата, в котором идет общение')
     text = models.TextField(null=True, blank=True, verbose_name='Текст сообщения')
     json = models.TextField()
     mailing = models.ForeignKey(Mailing, related_name='messages', on_delete=models.PROTECT, blank=True, null=True)
@@ -146,7 +146,7 @@ class CallbackData(models.Model):
 
     date = models.DateTimeField(null=True, verbose_name='Дата отправки')
     call_id = models.CharField(max_length=500, verbose_name='Идентификатор данных')
-    chat_id = models.IntegerField(verbose_name='Идентификатор чата из которого пришли данные')
+    chat_id = models.BigIntegerField(verbose_name='Идентификатор чата из которого пришли данные')
     text = models.TextField(null=True, blank=True, verbose_name='Текст сообщения')
     json = models.TextField()
